backend/requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
langchain
langchain-openai
langchain-community
openai
faiss-cpu
tiktoken
